+++ trunk/com.bookworm.main.MainActivity/src/com/bookworm/model/Comment.java
@@ -9,18 +9,18 @@ import java.util.Date;
  */
 public class Comment implements java.io.Serializable {
 
-	private Integer commentId;
-	private Integer commenterId;
+	private Long commentId;
+	private Long commenterId;
 	private String commentText;
-	private Integer commentedBookId;
-	private Integer commentedBookAdderId;
+	private Long commentedBookId;
+	private Long commentedBookAdderId;
 	private Date creationDate;
 
 	public Comment() {
 	}
 
-	public Comment(Integer commenterId, String commentText,
-			Integer commentedBookId, Integer commentedBookAdderId,
+	public Comment(Long commenterId, String commentText,
+			Long commentedBookId, Long commentedBookAdderId,
 			Date creationDate) {
 		this.commenterId = commenterId;
 		this.commentText = commentText;
@@ -29,19 +29,19 @@ public class Comment implements java.io.Serializable {
 		this.creationDate = creationDate;
 	}
 
-	public Integer getCommentId() {
+	public Long getCommentId() {
 		return this.commentId;
 	}
 
-	public void setCommentId(Integer commentId) {
+	public void setCommentId(Long commentId) {
 		this.commentId = commentId;
 	}
 
-	public Integer getCommenterId() {
+	public Long getCommenterId() {
 		return this.commenterId;
 	}
 
-	public void setCommenterId(Integer commenterId) {
+	public void setCommenterId(Long commenterId) {
 		this.commenterId = commenterId;
 	}
 
@@ -53,19 +53,19 @@ public class Comment implements java.io.Serializable {
 		this.commentText = commentText;
 	}
 
-	public Integer getCommentedBookId() {
+	public Long getCommentedBookId() {
 		return this.commentedBookId;
 	}
 
-	public void setCommentedBookId(Integer commentedBookId) {
+	public void setCommentedBookId(Long commentedBookId) {
 		this.commentedBookId = commentedBookId;
 	}
 
-	public Integer getCommentedBookAdderId() {
+	public Long getCommentedBookAdderId() {
 		return this.commentedBookAdderId;
 	}
 
-	public void setCommentedBookAdderId(Integer commentedBookAdderId) {
+	public void setCommentedBookAdderId(Long commentedBookAdderId) {
 		this.commentedBookAdderId = commentedBookAdderId;
 	}
